refactoring code

+++ models/entity.student.go
@@ -5,7 +5,7 @@ import (
 )
 
 type EntityStudent struct {
-	ID        uint   `gorm:"type:bigserial;primaryKey;autoIncrement"`
+	ID        uint64 `gorm:"type:bigserial;primaryKey;autoIncrement"`
 	Name      string `gorm:"type:varchar(255);not null"`
 	Npm       int    `gorm:"type:bigint;unique;not null"`
 	Fak       string `gorm:"type:varchar(255);not null"`
